LPS-199033 LPS-201740 depot-test: Refactor test
